feat: export getSlugEntries
