+++ src/lib/modules/parsnip/index.ts
@@ -1,3 +1,4 @@
 export { menuPageServerLoad } from "./route-util/menuPageServerLoad";
 export { slugPageServerLoad } from "./route-util/slugPageServerLoad";
+export { getSlugEntries } from "./route-util/getSlugEntries";
 export { default as ParsnipBlog } from "./route-util/ParsnipBlog.svelte";
